// part9/9.21-9.30/patientor-backend/src/utils.ts
import { z } from 'zod';
import { Diagnosis, Entry, Gender, HealthCheckRating } from './types';
import { NextFunction, Request, Response } from 'express';

const parseDiagnosisCodes = (object: unknown): Array<Diagnosis['code']> => {
  if (!object || typeof object !== 'object' || !('diagnosisCodes' in object)) {
    // we will just trust the data to be in correct form
    return [] as Array<Diagnosis['code']>;
  }

  return object.diagnosisCodes as Array<Diagnosis['code']>;
};

export const NewPatientSchema = z.object({
  name: z.string(),
  dateOfBirth: z.string().date(),
  ssn: z.string(),
  gender: z.nativeEnum(Gender),
  occupation: z.string(),
  entries: z.array(z.custom<Entry>()).default([]),
});

export const NewBaseEntrySchema = z.object({
  description: z.string().nonempty(),
  date: z.string().date().nonempty(),
  specialist: z.string().nonempty(),
  diagnosisCodes: z
    .unknown()
    .transform((val) => parseDiagnosisCodes({ diagnosisCodes: val })),
});

const NewHealthCheckSchema = NewBaseEntrySchema.extend({
  type: z.literal('HealthCheck'),
  healthCheckRating: z.nativeEnum(HealthCheckRating),
});

const NewOccupationalHealthcareSchema = NewBaseEntrySchema.extend({
  type: z.literal('OccupationalHealthcare'),
  employerName: z.string(),
  sickLeave: z
    .object({
      startDate: z.string().date(),
      endDate: z.string().date(),
    })
    .optional(),
});

const NewHospitalSchema = NewBaseEntrySchema.extend({
  type: z.literal('Hospital'),
  discharge: z.object({
    date: z.string().date(),
    criteria: z.string(),
  }),
});

export const NewEntrySchema = z.union([
  NewHealthCheckSchema,
  NewOccupationalHealthcareSchema,
  NewHospitalSchema,
]);

export const newPatientParser = (
  req: Request,
  _res: Response,
  next: NextFunction
) => {
  try {
    NewPatientSchema.parse(req.body);
    next();
  } catch (error: unknown) {
    next(error);
  }
};

export const newEntryParser = (
  req: Request,
  _res: Response,
  next: NextFunction
) => {
  try {
    console.log(req.body);
    NewEntrySchema.parse(req.body);
    next();
  } catch (error: unknown) {
    next(error);
  }
};

export const errorMiddleware = (
  error: unknown,
  _req: Request,
  res: Response,
  next: NextFunction
) => {
  if (error instanceof z.ZodError) {
    res.status(400).send(error.issues);
  } else {
    next(error);
  }
};
